requirements.txt
langchain
streamlit_chat
streamlit
python-dotenv
openai
os
